Add files via upload
This involves a simple search application (still working to enhance the UI and UX).  In this, you will have to add the basic information required for adding it to the database which could be used for retrieving as well.
By this one have their customized search results and do the search actions in a secured way.